Add files via upload
The Vowel Counter project is a web-based tool developed using HTML, CSS, and JavaScript. It allows users to input text and instantly calculates the number of vowels present within the text. With a simple and intuitive user interface, users can conveniently analyze textual data and obtain accurate vowel counts. This project demonstrates the power of front-end web technologies in providing useful tools for text analysis tasks. Contributions and feedback are appreciated to further enhance the functionality and usability of the Vowel Counter.